Some general fixes and changes (#59)
* General fixes and changes

- fixing typos
- fixing random capitals
- making some things more readable

* Replacing all "The `<x>`` command [...]" with "This command [...]"

The format for introduction was different between each file so I made them consistent.
Also means if commands change or similar it is one less thing to be updated thats in some files and not others etc.

* Add my own name (Kaiyumari) to the special thanks page

Added to the special thanks page in support/special-thanks.md, #Documentation Writers
Keeping the alphabetical order of the names

* Update moderation-commands/unmute.md

Following review by MoonlightCapital

Co-authored-by: MoonlightCapital <19293406+MoonlightCapital@users.noreply.github.com>

---------

Co-authored-by: MoonlightCapital <19293406+MoonlightCapital@users.noreply.github.com>